+++ src/app/[locale]/project/ProjectItem.tsx
