+++ src/app/[locale]/project/clientComponent.tsx
