+++ src/app/[locale]/project/index.module.css
