+++ src/messages/en.json
@@ -1,0 +1,18 @@
+{
+  "index": {
+    "greeting": "Hi,",
+    "title": "Welcome to PORTFOLIO",
+    "desc": "and I'm FrondEnd Developer"
+  },
+  "not-found": {
+    "title": "404 Not Found",
+    "content": "Could not find requested resource"
+  },
+  "about": {
+    "reason": {
+      "title": "🥰 WHY FRONTEND Developer?? 👨🏻‍💻",
+      "desc": "I was fascinated on software develop because developing process doesn't end up with data anaylsis, but also create my own product. I studied <a>JavaScript</a> to make web application and due to trend variation, I saw a lot of parts of dev parts converted to frontend. I choose frontend because I can see my code's changes result directly. Eventually, I work as frontend developer. I use <br /><a>react</a> in team project. I prioritize collaboration and try to challenge ourselves from various perspectives before waiting to solve problems. I feel proud when my team members use the code or results I wrote and then give feedback. So, we used swagger so that our team members could test the API right within the development environment. In workspace, I have used <a>TypeScript</a>, <a>next</a>, <a>sveltekit</a>."
+    },
+    "record": "Not only study, but also digest!!🧽 One practice is better than hundreds of study. I record after practice programming lauguages, libraries and frameworks to reuse after. I will write carefully so someone who enter my blog can study easier."
+  }
+}
