+++ src/messages/ko.json
@@ -1,0 +1,18 @@
+{
+  "index": {
+    "greeting": "안녕하세요,",
+    "title": "프로폴리오에 오신 것을 환영합니다.",
+    "desc": "저는 프론트엔드 개발자입니다.  "
+  },
+  "not-found": {
+    "title": "404 Not Found",
+    "content": "요청한 자원을 찾을 수 없습니다."
+  },
+  "about": {
+    "reason": {
+      "title": "🥰 프론트엔드 개발자의 길로...👨🏻‍💻",
+      "desc": "데이터 분석 공부하던 중 분석에서 끝나는 것이 아닌 본인의 프로덕트를 만들어 낼 수 있는 SW개발에 흥미가 생기게 되어 입문하게 되었습니다. <a>JavaScript</a> 기본기를 다지면서 개발 트렌드를 살펴보았고 트렌드의 변화로 인해 개발과정의 많은 부분이 프론트엔드로 넘어가는 것을 보았습니다. 코드 작성 후 결과물의 즉각적인 변화를 확인할 수 있는 점이 흥미에도 맞다고 생각하여 <strong>프론트엔드 직무</strong>로 나아가서 현재 주니어 개발자로 근무하고 있습니다. <br /><a>react</a> 기반 팀 프로젝트 경험이 있습니다. 협업을 우선시하며 문제 해결을 위해서는 기다리기보다 먼저 다양한 관점에서 도전하려고 합니다. 제가 작성한 코드나 결과물을 팀원이 사용하고나서 피드백할 때 뿌듯함을 느낍니다. 그래서 개발 환경 내에서 팀원들이 API를 바로 테스트 할 수 있게 swagger를 사용했습니다. 업무에서는 <a>TypeScript</a>, <a>next</a>, <a>sveltekit</a>를 사용해보았습니다."
+    },
+    "record": "배운 채로 방치하지 말고 내가 쓸 수 있도록 소화!🧽 적자생존! 적어야\n산다는 말이 있습니다. 프로그래밍 언어, 라이브러리 및 프레임워크를\n공부하고 사용하면서 나중에 재사용할만한 내용들을 수시로 블로그에\n기록하고 있습니다. 지나가던 사람이 제 블로그의 글을 읽는 것만으로 도움이\n되도록 열심히 적어보겠습니다."
+  }
+}
